Define abstract AffWP_Metabox_Base class
This includes a complete rewrite of reports metaboxes in AffiliateWP.

A metabox can be defined by extending the base class in includes/admin/class-metabox-base.php.

See: includes/admin/class-metabox-base.phpL28 for a full example

Note: Subsequent commits will define the $screen->id, accounting for all screen ID's in use by AffiliateWP. As it is now, any extension of this base class will be confined to the Reports page.